Allow customized subnet list for nmap
The core purpose for this change is to allow Bjorn to function in large
networks; 10.0.0.0/8 subnets have millions of available addresses. This
is a known limitation of nmap; nothing is well-suited to poll this many
hosts.

This commit adds two new config items, boolean 'use_custom_subnets' and
string[] 'custom_subnet_list'. By default, 'use_custom_subnets' is False
which matches the current behavior of using the IP Interface subnet. If
'use_custom_subnets' is True, then the subnets in 'custom_subnet_list'
are passed to nmap instead.